// frontend/src/app/shared/models/enums.model.ts
export enum ClothingCategory {
  TSHIRT = 'T-Shirt',
  HOODIE = 'Hoodie',
  SWEATER = 'Sweater',
  JACKET = 'Jacket',
  POLO = 'Polo',
  PANTS = 'Pants',
  SHORTS = 'Shorts',
  ACCESSORY = 'Accessory',
}

export enum Gender {
  MALE = 'Male',
  FEMALE = 'Female',
  UNISEX = 'Unisex',
}

export enum Size {
  XS = 'XS',
  S = 'S',
  M = 'M',
  L = 'L',
  XL = 'XL',
  XXL = 'XXL',
}

export enum Color {
  WHITE = 'White',
  BLACK = 'Black',
  RED = 'Red',
  BLUE = 'Blue',
  GREEN = 'Green',
  YELLOW = 'Yellow',
  PINK = 'Pink',
  GRAY = 'Gray',
  PURPLE = 'Purple',
  ORANGE = 'Orange',
}